_archived_noise/legacy_code/requirements.txt
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.20.0
scikit-learn>=1.0.0
numpy>=1.21.0
matplotlib>=3.5.0
textblob>=0.17.0
# Optional for SMS notifications
twilio>=9.0.0
# Scheduler
schedule>=1.2.0
# Alpaca paper/live trading
alpaca-py>=0.18.0
# Phase 8 dashboard
streamlit>=1.32.0
